flask_api/requirements.txt
#install flask

flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-socketio
opencv-python
ultralytics==8.3.55
pytesseract
numpy
flask-migrate
python-dotenv
requests
Werkzeug
supabase
flask-jwt-extended
flask-sqlalchemy
paddleocr
paddlepaddle
psycopg2-binary

